# docs/api/parent.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Parent API",
    "version": "2026-09-06"
  },
  "paths": {
...
    "/api/v1/parent/children/{child_id}/message/list": {
      "get": {
        "tags": [
          "父母授权"
        ],
        "summary": "父母查看已获同意的子女会话",
        "operationId": "conversations_api_v1_parent_children__child_id__message_list_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "child_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Child Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 9,
              "default": "",
              "title": "Cursor"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Pagesize"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageConversationPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/parent/children/{child_id}/message/applications": {
      "get": {
        "tags": [
          "父母授权"
        ],
        "summary": "父母查看子女申请分页",
        "operationId": "applications_api_v1_parent_children__child_id__message_applications_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "child_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Child Id"
            }
          },
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(incoming|outgoing)$",
              "default": "incoming",
              "title": "Direction"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 9,
              "default": "",
              "title": "Cursor"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Pagesize"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageApplicationPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/parent/children/{child_id}/message/chat/permission": {
      "get": {
        "tags": [
          "父母授权"
        ],
        "summary": "父母聊天前核验双方同意",
        "operationId": "permission_api_v1_parent_children__child_id__message_chat_permission_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "child_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Child Id"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Userid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatPermissionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/parent/children/{child_id}/message/chat": {
      "get": {
        "tags": [
          "父母授权"
        ],
        "summary": "父母读取脱敏聊天记录",
        "operationId": "chat_api_v1_parent_children__child_id__message_chat_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "child_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Child Id"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Userid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageChatItem"
                  },
                  "title": "Response Chat Api V1 Parent Children  Child Id  Message Chat Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/parent/children/{child_id}/message/send": {
      "post": {
        "tags": [
          "父母授权"
        ],
        "summary": "父母发送已获授权的文字消息",
        "operationId": "send_api_v1_parent_children__child_id__message_send_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "child_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Child Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageSendRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageSendResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/parent/children/{child_id}/message/application/handle": {
      "post": {
        "tags": [
          "父母授权"
        ],
        "summary": "父母处理子女收到的申请",
        "operationId": "handle_api_v1_parent_children__child_id__message_application_handle_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "child_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Child Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationHandleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationHandleResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/parent/children/{child_id}/message/read-all": {
      "post": {
        "tags": [
          "父母授权"
        ],
        "summary": "父母标记子女会话已读",
        "operationId": "read_all_api_v1_parent_children__child_id__message_read_all_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "child_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Child Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkAllReadResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/parent/children/{child_id}/message/messages/{message_id}": {
      "delete": {
        "tags": [
          "父母授权"
        ],
        "summary": "父母撤回授权主体发出的消息",
        "operationId": "revoke_message_api_v1_parent_children__child_id__message_messages__message_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "child_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Child Id"
            }
          },
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Message Id"
            }
          },
          {
            "name": "clientCommandId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "title": "Clientcommandid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageRevokeResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "MessageConversationPage": {
        "properties": {
          "list": {
            "items": {
              "$ref": "#/components/schemas/MessageConversation"
            },
            "type": "array",
            "title": "List"
          },
          "nextCursor": {
            "type": "string",
            "title": "Nextcursor"
          },
          "hasMore": {
            "type": "boolean",
            "title": "Hasmore"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "unreadTotal": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Unreadtotal"
          }
        },
        "type": "object",
        "required": [
          "list",
          "nextCursor",
          "hasMore",
          "total",
          "unreadTotal"
        ],
        "title": "MessageConversationPage"
      },
      "MessageConversation": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "conversationId": {
            "type": "string",
            "title": "Conversationid"
          },
          "userId": {
            "type": "integer",
            "title": "Userid"
          },
          "avatar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "lastMessage": {
            "type": "string",
            "title": "Lastmessage"
          },
          "time": {
            "type": "integer",
            "title": "Time"
          },
          "unreadCount": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Unreadcount"
          },
          "online": {
            "type": "boolean",
            "title": "Online",
            "default": false
          },
          "messageType": {
            "type": "string",
            "enum": [
              "text",
              "image",
              "voice",
              "video"
            ],
            "title": "Messagetype",
            "default": "text"
          },
          "type": {
            "type": "string",
            "const": "chat",
            "title": "Type",
            "default": "chat"
          },
          "matched": {
            "type": "boolean",
            "title": "Matched",
            "default": true
          },
          "canChat": {
            "type": "boolean",
            "title": "Canchat",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "conversationId",
          "userId",
          "avatar",
          "name",
          "lastMessage",
          "time",
          "unreadCount"
        ],
        "title": "MessageConversation"
      },
      "MessageApplicationPage": {
        "properties": {
          "list": {
            "items": {
              "$ref": "#/components/schemas/MessageApplication"
            },
            "type": "array",
            "title": "List"
          },
          "nextCursor": {
            "type": "string",
            "title": "Nextcursor"
          },
          "hasMore": {
            "type": "boolean",
            "title": "Hasmore"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "pendingCount": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pendingcount",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "list",
          "nextCursor",
          "hasMore",
          "total"
        ],
        "title": "MessageApplicationPage"
      },
      "MessageApplication": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "userId": {
            "type": "integer",
            "title": "Userid"
          },
          "avatar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "time": {
            "type": "integer",
            "title": "Time"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "accepted",
              "rejected",
              "expired"
            ],
            "title": "Status"
          },
          "statusText": {
            "type": "string",
            "title": "Statustext"
          },
          "direction": {
            "type": "string",
            "enum": [
              "in",
              "out"
            ],
            "title": "Direction"
          }
        },
        "type": "object",
        "required": [
          "id",
          "userId",
          "avatar",
          "name",
          "message",
          "time",
          "status",
          "statusText",
          "direction"
        ],
        "title": "MessageApplication"
      },
      "ChatPermissionResponse": {
        "properties": {
          "userId": {
            "type": "integer",
            "title": "Userid"
          },
          "conversationId": {
            "type": "string",
            "title": "Conversationid",
            "default": ""
          },
          "sessionId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sessionid"
          },
          "canChat": {
            "type": "boolean",
            "title": "Canchat"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "userId",
          "canChat",
          "reason"
        ],
        "title": "ChatPermissionResponse"
      },
      "MessageChatItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "clientMessageId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clientmessageid"
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "image",
              "voice",
              "video"
            ],
            "title": "Type"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "time": {
            "type": "integer",
            "title": "Time"
          },
          "isMine": {
            "type": "boolean",
            "title": "Ismine"
          },
          "senderAvatar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Senderavatar"
          },
          "revoked": {
            "type": "boolean",
            "title": "Revoked",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "content",
          "time",
          "isMine"
        ],
        "title": "MessageChatItem"
      },
      "MessageSendRequest": {
        "properties": {
          "userId": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Userid"
          },
          "content": {
            "type": "string",
            "maxLength": 5000,
            "title": "Content",
            "default": ""
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "image",
              "voice",
              "video"
            ],
            "title": "Type",
            "default": "text"
          },
          "clientMessageId": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Clientmessageid"
          },
          "mediaId": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Mediaid"
          }
        },
        "type": "object",
        "required": [
          "userId",
          "clientMessageId"
        ],
        "title": "MessageSendRequest"
      },
      "MessageSendResult": {
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "title": "Success",
            "default": true
          },
          "messageId": {
            "type": "integer",
            "title": "Messageid"
          },
          "message": {
            "$ref": "#/components/schemas/MessageChatItem"
          },
          "deduplicated": {
            "type": "boolean",
            "title": "Deduplicated",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "messageId",
          "message"
        ],
        "title": "MessageSendResult"
      },
      "ApplicationHandleRequest": {
        "properties": {
          "applicationId": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Applicationid"
          },
          "action": {
            "type": "string",
            "enum": [
              "accept",
              "reject"
            ],
            "title": "Action"
          },
          "clientCommandId": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Clientcommandid"
          }
        },
        "type": "object",
        "required": [
          "applicationId",
          "action",
          "clientCommandId"
        ],
        "title": "ApplicationHandleRequest"
      },
      "ApplicationHandleResult": {
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "title": "Success",
            "default": true
          },
          "application": {
            "$ref": "#/components/schemas/MessageApplication"
          },
          "canChat": {
            "type": "boolean",
            "title": "Canchat"
          }
        },
        "type": "object",
        "required": [
          "application",
          "canChat"
        ],
        "title": "ApplicationHandleResult"
      },
      "MarkAllReadResult": {
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "title": "Success",
            "default": true
          },
          "updatedCount": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Updatedcount"
          },
          "unreadTotal": {
            "type": "integer",
            "const": 0,
            "title": "Unreadtotal",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "updatedCount"
        ],
        "title": "MarkAllReadResult"
      },
      "MessageRevokeResult": {
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "title": "Success",
            "default": true
          },
          "messageId": {
            "type": "integer",
            "title": "Messageid"
          }
        },
        "type": "object",
        "required": [
          "messageId"
        ],
        "title": "MessageRevokeResult"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}